// crtr_angular/src/app/solicitacao/nova-solicitacao/cadastro-procuracoes/models/outorgante-existente.model.ts
export class OutorganteExistente{
  idProcuracaoParte: number;
  idSolicitacao: number;
  idPessoa: number;
  nome: string;
  idTipoDocumento: number;
  documento: number;
  idTipoProcuracaoParte: number;
  idProcuracaoParteEstado: number;
  jsonConteudo: string;
  conteudoPessoasFisicas: string;
  conteudoPessoasContatos: string;
  enderecoEntrega: string;
  email: string;
  idUsuario: number;
}