// crtr_angular/src/app/solicitacao/nova-solicitacao/cadastro-procuracoes/models/outorgado-existente.model.ts
export class OutorgadoExistente{
  idProcuracaoParte: number;
  idSolicitacao: number;
  idUsuario: number;
  idPessoa: number;
  nome: string;
  email: string;
  documento: number;
  enderecoEntrega: string
  idTipoProcuracaoParte: number;
  idTipoDocumento: number;
  idProcuracaoParteEstado: number;
  jsonConteudo: string;
  conteudoPessoasFisicas: string;
  conteudoPessoasContatos: string;
  dataOperacao: string;
}